renamed z files + added class template for serverz
