+++ src/serverz.py
@@ -1,0 +1,12 @@
+# @Author: George Onoufriou <georgeraven>
+# @Date:   2018-11-05
+# @Filename: serverz.py
+# @Last modified by:   georgeraven
+# @Last modified time: 2018-11-05
+# @License: Please see LICENSE in project root.
+# @Copyright: George Onoufriou
+
+class serverz():
+
+    def __init__(self):
+        None
